requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
transformers>=4.45.0
torch>=2.1.0
torchvision>=0.16.0
accelerate>=0.26.0
Pillow>=10.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
groq>=0.9.0
tavily-python>=0.3.0
datasets>=2.20.0
tqdm>=4.66.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sentencepiece>=0.2.0
qwen-vl-utils>=0.0.8
spacy>=3.7.0
